Feature michiel

+++ Rekenmachine1/src/Michiel.java
@@ -1,8 +1,8 @@
 
 public class Michiel {
 	
-	public int delen(int x, int y) {
-		int a = x / y;
+	public double delen(double x, double y) {
+		double a = x / y;
 		return a;}
 	
 	public int vermenigvuldigen(int x, int y) {
